refactor decorator scanning
