+++ src/utils/decorators/decoratorFactory.ts
@@ -1,14 +1,8 @@
 import { UserError } from "../../errors.js";
-import FileDetails from "../../models/fileDetails.js";
 import { Language } from "../../yamlProjectConfiguration/models.js";
-import { ClassInfo } from "./decoratorTypes.js";
+import { DecoratorExtractor } from "./baseDecoratorExtractor.js";
 import { GoDecoratorExtractor } from "./goDecorators.js";
 import { JsTsDecoratorExtractor } from "./jsTsDecorators.js";
-
-export interface DecoratorExtractor {
-    getDecoratorsFromFile: (file: string) => Promise<ClassInfo[]>;
-    fileFilter: (cwd: string) => (file: FileDetails) => boolean;
-}
 
 export class DecoratorExtractorFactory {
     static createExtractor = (language: Language): DecoratorExtractor => {
